线程类的封装初步可以，但是需要把loop函数编程虚函数从而子类从基类继承派生

+++ libraries/Bit_scheduler.h
@@ -9,13 +9,6 @@
 #ifndef LIBRARIES_BIT_SCHEDULER_H_
 #define LIBRARIES_BIT_SCHEDULER_H_
 
-/*
- * 这是任务调度
- * 在开始的时候要调用scheduler.init()，然后每次循环都调用scheduler.tick()，每次循环时间大概是10ms
- * 最后调用scheduler.run()，这个run函数中的参数时最大允许执行时间，也就是当这个允许的时间到了后，
- * 就需要返回了，不再执行任务
- */
-
 #include <stdint.h>
 
 class BIT_Scheduler
@@ -26,30 +19,23 @@ public:
     struct Task
     {
         task_fn_t function;
-        //uint16_t  rate_hz;
-        float     rate_hz;
-        //uint16_t  max_time_micros; // 目前用uint16_t表示，也就是最多是65毫秒，如果串口波特率是9600的，这个指的是传输速率，并不是读取速率，读取速率比传输速率要快很多
+        double     rate_hz;
         uint32_t  max_time_micros;
     };
 
     // 调度初始化，指定任务表的地址，明确有多少个任务
     void init(const Task *tasks, uint8_t num_tasks);
 
-    // 每一次tick都会把计数加1
+    // 主程序是个循环，每次循环称为一次tick，tick()函数每执行一次会把_tick_counter计数加1
     void tick(void);
 
-    /*
-     * 每一次tick都会调用一次这个run函数去执行尚未完成的scheduler中的任务
-     * 执行调度任务表中的任务，参数是最大允许时间
-     * 时间单位是微秒
-     * 这个会在最大允许时间内执行尽可能多的任务
-     */
+    // 在time_available时间内执行任务调度表尽可能多的任务
     void run(uint32_t time_available);
 
     // return the number of microseconds available for the current task
     uint32_t time_available_usec(void);
 
-    uint16_t get_loop_rate_hz(void); // 目前设置频率范围是 1~1000hz
+    uint16_t get_loop_rate_hz(void);
 
 private:
     // progmem list of tasks to run
@@ -58,8 +44,8 @@ private:
     // number of tasks in _tasks list
     uint8_t _num_tasks;
 
-    // number of 'ticks' that have passed (number of times that
-    // tick() has been called
+    // number of 'ticks' that have passed
+    // number of times that tick() has been called
     uint16_t _tick_counter;
 
     // tick counter at the time we last ran each task
